Refactor Generator out of Encounter
