+++ src/space/encounter.js
@@ -7,6 +7,7 @@ import PlayerShip from "./objects/realization/playerShip";
 import AsteroidCreator from "./objects/realization/asteroid";
 import Barrel from "./objects/realization/barrel";
 import Duck from "./objects/realization/duck";
+import Generator from "./generator";
 
 /** Name of the image used for the background */
 const BACKGROUND_IMAGE = 'spaceBackground/starBackground';
@@ -41,102 +42,10 @@ export default class Encounter {
         this.height = Math.max(height, game.HEIGHT);
 
         this.camera = new Camera(new Vector(this.width, this.height), new Vector(game.WIDTH, game.HEIGHT));
-        this.prepareBackground();
-        this.initialize();
-    }
+        this.generator = new Generator(this);
 
-    /**
-     * Initializes the encounter
-     */
-    initialize() {
-        // TODO: Prepare game objects and win conditions
-
-        this.playerShip = new PlayerShip(this, new Vector(this.width / 2, this.height / 2));
-        this.addObject(this.playerShip);
-
-        this.camera.bindTo(this.playerShip);
-
-        // AsteroidCreator.createLineAsteroidShower(
-        //     this,
-        //     new Vector(this.width / 2 - 350, this.height / 2 + 100),
-        //     20,
-        //     90,
-        //     true
-        //     ).forEach(asteroid => this.addObject(asteroid));
-        //
-        //
-        // let barrelSpacing = 80;
-        // for (let i = 0; i < 5; i++) {
-        //     this.addObject(new Barrel(this, new Vector(this.width / 2 - 2 * barrelSpacing + i * barrelSpacing, this.height / 2 + 2 * barrelSpacing)))
-        // }
-
-        let spawnPos = new Vector(this.width / 2 - 600, this.height / 2 - 30);
-        for (let x = 0; x < 10; x ++) {
-            for (let y = 0; y < 2; y++) {
-                this.addObject(new Duck(this, new Vector(spawnPos.x + x * 30, spawnPos.y + y * 30)))
-            }
-        }
-    }
-
-    /* private */ createLayer() {
-        let buffer = document.createElement('canvas');
-        buffer.width = this.width;
-        buffer.height = this.height;
-
-        return buffer;
-    }
-
-    /**
-     * Prepares the background image
-     */
-    prepareBackground() {
-        let buffer = this.createLayer();
-        let ctx = buffer.getContext('2d');
-
-        let backgroundImage = AssetLoader.getAsset(BACKGROUND_IMAGE);
-        let starSmallImage = AssetLoader.getAsset(STAR_SMALL_IMAGE);
-        let starBigImage = AssetLoader.getAsset(STAR_BIG_IMAGE);
-        let imageWidth = backgroundImage.width;
-        let imageHeight = backgroundImage.height;
-
-        for (let x = 0; x < this.width; x += imageWidth) {
-            for (let y = 0; y < this.height; y += imageHeight) {
-                ctx.drawImage(backgroundImage, x, y, imageWidth, imageHeight);
-            }
-        }
-
-
-        this.camera.addLayer(buffer, 0.2, new Vector(0, 0));
-        for (let i = 0; i < 7; i ++) {
-            let starBuffer = this.createLayer();
-            let starCtx = starBuffer.getContext('2d');
-            starCtx.globalAlpha = 0.7;
-
-            let amount = Math.sqrt(this.width * this.height) / 40;
-            for (let j = 0; j < amount; j++) {
-                let star = starSmallImage;
-                if (Math.random() < 0.5) star = starBigImage;
-                let x = Math.random() * this.width;
-                let y = Math.random() * this.height;
-
-                starCtx.drawImage(star, x, y);
-            }
-
-            if (i < 6) {
-                starCtx.globalAlpha = 1;
-                if (Math.random() < 0.5) {
-                    let planet = AssetLoader.getAsset(PLANET_IMAGE_PREFIX + (Math.floor(Math.random() * 12) + 1));
-                    let size = Math.random() * 500 + 250;
-                    let x = Math.random() * this.width;
-                    let y = Math.random() * this.height;
-
-                    starCtx.drawImage(planet, x - size / 2, y - size / 2, size, size);
-                }
-            }
-
-
-            this.camera.addLayer(starBuffer, 0.5 + i / 12.0, new Vector(0, 0));
-        }
+        this.generator.prepareBackground();
+        this.generator.initialize();
     }
 
     /**
@@ -171,22 +80,16 @@ export default class Encounter {
     }
 
     /**
-     * Callback function for objects to report they died
-     *
-     * @param {EncounterObject} object - game object, which got destroyed
-     */
-    onObjectDestroyed(object) {
-        // TODO: Check for win conditions
-    }
-
-    /**
      * Removes given object from game
      *
      * @param {EncounterObject} object - game object to be removed from game
      */
     removeObject(object) {
         // Postpone removing of objects to make sure we are not accessing gameObjects array while iterating trough it
-        if (object !== undefined && object !== null) this.markedForDeletion.push(object);
+        if (object !== undefined && object !== null) {
+            this.markedForDeletion.push(object);
+            this.generator.onObjectDestroyed(object)
+        }
     }
 
     /**
@@ -230,6 +133,8 @@ export default class Encounter {
 
         });
         this.handleCollisions();
+
+        this.generator.update(elapsedTime);
 
         this.camera.update(elapsedTime);
